Adding .Array to .sync and .future to allow async functions to return multiple values
For example, the request module/function passes 2 arguments
(in addition to err) to its callback -- response and body

If we synchronize/futureize it, then we only get access to the first
argument (response). But if we transform the function into one that
only passes one arg to its callback -- an Array containing everything
we want to return -- then we can synchronize/futureize it. (In this
particular example, response actually contains response.body so the
transform isn't necessary, but it makes for an easy-to-understand example)

This change lets you access that synchronized/futureized transformed
function through .sync.Array and .future.Array

    coffee> request = require 'request';
    coffee> [response, body] = request.sync.Array "http://ccheever.com/_ip.php"; body
    '24.6.54.10'

This was requested by someone else here as well:
https://github.com/goodeggs/fibrous/issues/17